third_party/WebKit/LayoutTests/external/wpt/html/browsers/history/the-history-interface/007-expected.txt

This is a testharness.js-based test.
PASS history.state should initially be null
PASS history.pushState support is needed for this testcase
PASS history.state should reflect pushed state
FAIL popstate event should fire before onload firesassert_true: expected true got false
FAIL the correct state should be restored when navigating during initial loadassert_equals: expected (string) "state1" but got (boolean) false
FAIL history.state should reflect the navigated state onloadassert_equals: expected "state1" but got "state2"
FAIL popstate event should not fire after onload firesassert_false: expected false got true
PASS history.state should reflect the navigated state after onload
Harness: the test ran to completion.
